Minor changes Added Warning flags and fixed warnings

+++ src/core/system.c
@@ -1,21 +1,21 @@
 #include "system.h"
 
 uint8_t *memcpy(uint8_t *dest, uint8_t *src, size_t count){
-	for(int i = 0; i < count; i++){
+	for(size_t i = 0; i < count; i++){
 		dest[i] = src[i];
 	}
 	return dest;
 }
 
 uint8_t *memset (uint8_t *dest, uint8_t val, size_t count){
-	for(int i = 0; i < count; i++){
+	for(size_t i = 0; i < count; i++){
 		dest[i] = val;
 	}
 	return dest;
 }
 
 uint16_t *memsetw (uint16_t *dest, uint16_t val, size_t count){
-	for(int i = 0; i < count; i++){
+	for(size_t i = 0; i < count; i++){
 		dest[i] = val;
 	}
 	return dest;
